requirements.txt
django
djangorestframework
django-filter
PyJWT
